requirements.txt
fastapi
numpy
pandas
protobuf
PyYAML
sentencepiece
tiktoken
tokenizers
transformers
uvicorn
cuda-python==12.9.2  #<13.0，根據本地cuda版本選擇